Rename `checker` module to `manifest`
