+++ src/manifest.py
