+++ tests/test_chromiumchecker.py
@@ -3,7 +3,7 @@ import os
 import unittest
 from distutils.version import LooseVersion
 
-from src.checker import ManifestChecker
+from src.manifest import ManifestChecker
 from src.lib.externaldata import (
     ExternalData,
     ExternalFile,
