+++ tests/test_debianrepochecker.py
@@ -2,7 +2,7 @@ import unittest
 import os
 import datetime
 
-from src.checker import ManifestChecker
+from src.manifest import ManifestChecker
 from src.lib.utils import init_logging
 from src.lib.checksums import MultiDigest
