+++ tests/test_gitchecker.py
@@ -2,7 +2,7 @@ import copy
 import os
 import unittest
 
-from src.checker import ManifestChecker
+from src.manifest import ManifestChecker
 from src.lib.externaldata import ExternalGitRepo, ExternalGitRef
 from src.lib.utils import init_logging
 from src.checkers.gitchecker import TagWithVersion, TagWithSemver
