+++ tests/test_rustchecker.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from src.checker import ManifestChecker
+from src.manifest import ManifestChecker
 from src.lib.utils import init_logging
 from src.lib.checksums import MultiDigest
